// src/app/common/enums/buttons.enum.ts
export enum Buttons {
  SAVE = 'WINK.BUTTONS.SAVE',
  CANCEL = 'WINK.BUTTONS.CANCEL',
  WINK = 'WINK.BUTTONS.WINK',
  IGNORE = 'WINK.BUTTONS.IGNORE',
  DELETE = 'WINK.BUTTONS.DELETE',
  OK = 'WINK.BUTTONS.OK',
  DISCARD = 'WINK.BUTTONS.DISCARD',
  YES = 'WINK.BUTTONS.YES',
  NO = 'WINK.BUTTONS.NO',
  END = 'WINK.BUTTONS.END',
  PREVIOUS = 'WINK.BUTTONS.PREVIOUS',
  PREV = 'WINK.BUTTONS.PREV',
  NEXT = 'WINK.BUTTONS.NEXT',
  NEXT2 = 'WINK.BUTTONS.NEXT2'
}