// src/app/common/enums/messageTour.enum.ts
export enum MessageTour {
  HOME_STATUS = 'WINK.TOUR.HOME.STATUS',
  HOME_SEE_STATUS = 'WINK.TOUR.HOME.SEE_STATUS',
  HOME_PROFILES = 'WINK.TOUR.HOME.PROFILES',
  HOME_CARD = 'WINK.TOUR.HOME.CARD',


  WINKS_INTRODUCTION = 'WINK.TOUR.WINKS.INTRODUCTION',
  WINKS_REQUESTS = 'WINK.TOUR.WINKS.REQUESTS',
  WINKS_WINK = 'WINK.TOUR.WINKS.WINK',
  WINKS_RECORD = 'WINK.TOUR.WINKS.RECORD',
  WINKS_DELETE = 'WINK.TOUR.WINKS.DELETE',


  SETTINGS_SECTIONS = 'WINK.TOUR.SETTINGS.SECTIONS',
  SETTINGS_PUBLIC = 'WINK.TOUR.SETTINGS.PUBLIC',
  SETTINGS_PRIVATE = 'WINK.TOUR.SETTINGS.PRIVATE',
  SETTINGS_ADD = 'WINK.TOUR.SETTINGS.ADD',
  SETTINGS_ORDER = 'WINK.TOUR.SETTINGS.ORDER',
  SETTINGS_MENU = 'WINK.TOUR.SETTINGS.MENU',
}